[refactor][x]: move reusable ErrorMessage component to _shared
