+++ components/_shared/index.ts
@@ -1,0 +1,4 @@
+import Table from './Table';
+import ErrorMessage from './Error';
+
+export { Table, ErrorMessage };
